chore(code): Move template implementations to separate _impl hpp for neatness
